Backend/requirements.txt
aiohttp==3.8.4
aiohttp-cors==0.7.0
aiohttp-devtools==1.0.post0
aioice==0.7.6
aiortc==1.4.0
aiosignal==1.3.1
anyio==3.6.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
av==10.0.0
boto3==1.26.92
botocore==1.29.92
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.2
devtools==0.10.0
dill==0.3.6
dnspython==2.3.0
executing==1.2.0
filelock==3.9.0
frozenlist==1.3.3
google-crc32c==1.5.0
huggingface-hub==0.13.2
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
MarkupSafe==2.1.2
mpmath==1.2.1
multidict==6.0.4
netifaces==0.11.0
networkx==3.0
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
Pillow==9.4.0
pycparser==2.21
pyee==9.0.4
Pygments==2.14.0
pylibsrtp==0.8.0
pyOpenSSL==23.0.0
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
s3transfer==0.6.0
sacremoses==0.0.53
sentencepiece==0.1.97
six==1.16.0
sniffio==1.3.0
sympy==1.11.1
tokenizers==0.13.2

tqdm==4.65.0
transformers==4.27.1
typing-extensions==4.5.0
urllib3==1.26.13
watchgod==0.8.2
yarl==1.8.2
